+++ packages/@vangware/notify/src/EventTypeDictionary.ts
@@ -3,6 +3,6 @@ import type { ReadOnlyRecord } from "@vangware/types";
 /**
  * Dictionary of event name to event types.
  *
- * @see [ReadOnlyRecord](https://libraries.vangware.com/types/_vangware_types.ReadOnlyRecord.html)
+ * @see [ReadOnlyRecord](https://vangware.com/libraries/vangware_types/#readonlyrecord)
  */
 export type EventTypeDictionary = ReadOnlyRecord<string>;