+++ packages/@vangware/ansi/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change log
 
+## 6.3.29
+
+### Patch Changes
+
+-   📝 update documentation and URLs.
+-   ⬆️ upgrade `@vangware/types` to `v6.2.10`
+
 ## 6.3.28
 
 ### Patch Changes